@@ -27,7 +27,7 @@
 
   <groupId>io.jenkins.pluginhealth.scoring</groupId>
   <artifactId>plugin-health-scoring-parent</artifactId>
-  <version>4.4.1-SNAPSHOT</version>
+  <version>4.10.2-SNAPSHOT</version>
   <packaging>pom</packaging>
 
   <name>Plugin Health Scoring :: Parent</name>
@@ -81,8 +81,8 @@
     <corepack.version>0.32.0</corepack.version>
 
     <jjwt.version>0.12.6</jjwt.version>
-    <testcontainers.version>1.20.6</testcontainers.version>
-    <spring-boot.version>3.4.3</spring-boot.version>
+    <testcontainers.version>1.21.1</testcontainers.version>
+    <spring-boot.version>3.5.0</spring-boot.version>
   </properties>
 
   <dependencyManagement>
@@ -140,7 +140,7 @@
       <dependency>
         <groupId>org.kohsuke</groupId>
         <artifactId>github-api</artifactId>
-        <version>2.0-rc.1</version>
+        <version>2.0-rc.3</version>
       </dependency>
     </dependencies>
   </dependencyManagement>
@@ -164,7 +164,7 @@
         <plugin>
           <groupId>com.diffplug.spotless</groupId>
           <artifactId>spotless-maven-plugin</artifactId>
-          <version>2.44.3</version>
+          <version>2.44.5</version>
           <configuration>
             <ratchetFrom>origin/main</ratchetFrom>
             <java>
@@ -223,7 +223,7 @@
             <dependency>
               <groupId>com.puppycrawl.tools</groupId>
               <artifactId>checkstyle</artifactId>
-              <version>10.21.4</version>
+              <version>10.25.0</version>
             </dependency>
           </dependencies>
         </plugin>
@@ -239,7 +239,7 @@
         </plugin>
         <plugin>
           <artifactId>maven-failsafe-plugin</artifactId>
-          <version>3.5.2</version>
+          <version>3.5.3</version>
           <configuration>
             <classesDirectory>${project.build.outputDirectory}</classesDirectory>
           </configuration>
@@ -263,12 +263,12 @@
         </plugin>
         <plugin>
           <artifactId>maven-surefire-plugin</artifactId>
-          <version>3.5.2</version>
+          <version>3.5.3</version>
         </plugin>
         <plugin>
           <groupId>org.jacoco</groupId>
           <artifactId>jacoco-maven-plugin</artifactId>
-          <version>0.8.12</version>
+          <version>0.8.13</version>
           <configuration>
             <formats>
               <format>XML</format>
