@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2023 Jenkins Infra
+ * Copyright (c) 2023-2025 Jenkins Infra
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,12 +21,11 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package io.jenkins.pluginhealth.scoring.service;
 
 import java.io.IOException;
 import java.net.MalformedURLException;
-import java.net.URL;
+import java.net.URI;
 import java.util.Map;
 import java.util.stream.Collectors;
 
@@ -53,15 +52,13 @@ public class PluginDocumentationService {
     public Map<String, String> fetchPluginDocumentationUrl() {
         try {
             final Map<String, Link> documentationUrlsMap = objectMapper.readValue(
-                new URL(configuration.jenkins().documentationUrls()),
-                  new TypeReference<>() {
-                  }
-            );
+                    URI.create(configuration.jenkins().documentationUrls()).toURL(), new TypeReference<>() {});
             return documentationUrlsMap.entrySet().stream()
-                .collect(Collectors.toMap(
-                    Map.Entry::getKey,
-                    e -> e.getValue() == null || e.getValue().url() == null ? "" : e.getValue().url()
-                ));
+                    .collect(Collectors.toMap(
+                            Map.Entry::getKey,
+                            e -> e.getValue() == null || e.getValue().url() == null
+                                    ? ""
+                                    : e.getValue().url()));
         } catch (MalformedURLException e) {
             LOGGER.error("URL to documentation link is incorrect.", e);
             return Map.of();
