@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2023-2024 Jenkins Infra
+ * Copyright (c) 2023-2025 Jenkins Infra
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -52,7 +52,7 @@ public class KnownSecurityVulnerabilityProbe extends Probe {
                     }
                     final Pattern pattern = Pattern.compile(securityWarningVersion.pattern());
                     final Matcher matcher = pattern.matcher(plugin.getVersion().toString());
-                    return matcher.find();
+                    return matcher.matches();
                 }))
                 .map(warning -> String.join("|", warning.id(), warning.url()))
                 .collect(Collectors.joining(", "));
