@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2023-2024 Jenkins Infra
+ * Copyright (c) 2023-2025 Jenkins Infra
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -265,6 +265,99 @@ class KnownSecurityVulnerabilityProbeTest extends AbstractProbeTest<KnownSecurit
                 .comparingOnlyFields("id", "status", "message")
                 .isEqualTo(ProbeResult.success(
                         KnownSecurityVulnerabilityProbe.KEY, warningId1 + "|" + url1, probe.getVersion()));
+    }
+
+    @Test
+    void swarm349ShouldNotHaveSecurityAdvisory() {
+        final String pluginName = "swarm";
+        final VersionNumber pluginVersion = new VersionNumber("3.49");
+        final String warningId = "SECURITY-597";
+        final String url = "https://jenkins.io/security/advisory/2017-10-11/";
+        final var plugin = mock(io.jenkins.pluginhealth.scoring.model.Plugin.class);
+        final ProbeContext ctx = mock(ProbeContext.class);
+        final KnownSecurityVulnerabilityProbe probe = getSpy();
+
+        when(plugin.getName()).thenReturn(pluginName);
+        when(plugin.getVersion()).thenReturn(pluginVersion);
+        when(ctx.getUpdateCenter())
+                .thenReturn(new UpdateCenter(
+                        Collections.emptyMap(),
+                        Collections.emptyMap(),
+                        List.of(new SecurityWarning(
+                                warningId,
+                                pluginName,
+                                url,
+                                List.of(new SecurityWarningVersion(null, "([12][.].+|3[.][34])(|[.-].*)"))))));
+
+        final ProbeResult result = probe.apply(plugin, ctx);
+        assertThat(result)
+                .isNotNull()
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.success(
+                        KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", probe.getVersion()));
+    }
+
+    @Test
+    void swarm35ShouldNotHaveSecurityAdvisory() {
+        final String pluginName = "swarm";
+        final VersionNumber pluginVersion = new VersionNumber("3.5");
+        final String warningId = "SECURITY-597";
+        final String url = "https://jenkins.io/security/advisory/2017-10-11/";
+        final var plugin = mock(io.jenkins.pluginhealth.scoring.model.Plugin.class);
+        final ProbeContext ctx = mock(ProbeContext.class);
+        final KnownSecurityVulnerabilityProbe probe = getSpy();
+
+        when(plugin.getName()).thenReturn(pluginName);
+        when(plugin.getVersion()).thenReturn(pluginVersion);
+        when(ctx.getUpdateCenter())
+                .thenReturn(new UpdateCenter(
+                        Collections.emptyMap(),
+                        Collections.emptyMap(),
+                        List.of(new SecurityWarning(
+                                warningId,
+                                pluginName,
+                                url,
+                                List.of(new SecurityWarningVersion(null, "([12][.].+|3[.][34])(|[.-].*)"))))));
+
+        final ProbeResult result = probe.apply(plugin, ctx);
+        assertThat(result)
+                .isNotNull()
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.success(
+                        KnownSecurityVulnerabilityProbe.KEY, "No known security vulnerabilities.", probe.getVersion()));
+    }
+
+    @Test
+    void swarm34ShouldHaveSecurityAdvisory() {
+        final String pluginName = "swarm";
+        final VersionNumber pluginVersion = new VersionNumber("3.4");
+        final String warningId = "SECURITY-597";
+        final String url = "https://jenkins.io/security/advisory/2017-10-11/";
+        final var plugin = mock(io.jenkins.pluginhealth.scoring.model.Plugin.class);
+        final ProbeContext ctx = mock(ProbeContext.class);
+        final KnownSecurityVulnerabilityProbe probe = getSpy();
+
+        when(plugin.getName()).thenReturn(pluginName);
+        when(plugin.getVersion()).thenReturn(pluginVersion);
+        when(ctx.getUpdateCenter())
+                .thenReturn(new UpdateCenter(
+                        Collections.emptyMap(),
+                        Collections.emptyMap(),
+                        List.of(new SecurityWarning(
+                                warningId,
+                                pluginName,
+                                url,
+                                List.of(new SecurityWarningVersion(null, "([12][.].+|3[.][34])(|[.-].*)"))))));
+
+        final ProbeResult result = probe.apply(plugin, ctx);
+        assertThat(result)
+                .isNotNull()
+                .usingRecursiveComparison()
+                .comparingOnlyFields("id", "status", "message")
+                .isEqualTo(ProbeResult.success(
+                        KnownSecurityVulnerabilityProbe.KEY, warningId + "|" + url, probe.getVersion()));
     }
 
     @Test
