@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2023 Jenkins Infra
+ * Copyright (c) 2023-2025 Jenkins Infra
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,18 +21,61 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
 package io.jenkins.pluginhealth.scoring.http;
 
+import java.util.List;
+import java.util.Map;
+
+import io.jenkins.pluginhealth.scoring.model.Plugin;
+import io.jenkins.pluginhealth.scoring.service.PluginService;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.ModelAndView;
 
 @Controller
 @RequestMapping(path = "/")
 public class IndexController {
+    private final PluginService pluginService;
+
+    public IndexController(PluginService pluginService) {
+        this.pluginService = pluginService;
+    }
+
+    @ModelAttribute(name = "module")
+    /* default */ String module() {
+        return "index";
+    }
+
     @GetMapping
     public String index() {
-        return "redirect:/probes";
+        return "index";
+    }
+
+    public record SearchRequest(String query) {
+        public SearchRequest {
+            query = query.trim();
+        }
+    }
+
+    @PostMapping(consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+    public ModelAndView search(@RequestParam SearchRequest search) {
+        final List<Plugin> results = pluginService.search(search.query());
+        if (results.isEmpty()) {
+            final ModelAndView modelAndView =
+                    new ModelAndView("index", Map.of("error", "No plugin found with query '" + search.query() + "'"));
+            modelAndView.setStatus(HttpStatus.NOT_FOUND);
+            return modelAndView;
+        } else if (results.size() == 1) {
+            return new ModelAndView("redirect:/scores/" + results.getFirst().getName());
+        } else {
+            return new ModelAndView("index", Map.of("results", results, "search", search.query()));
+        }
     }
 }
